transaction updated

+++ accountsService/src/main/java/com/retailBanking/accountsService/Repository/CreditCardRepository.java
@@ -9,9 +9,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.retailBanking.accountsService.Models.CreditCardModel;
 
-public interface CreditCardRepository extends JpaRepository<CreditCardModel, BigInteger>{
+public interface CreditCardRepository extends JpaRepository<CreditCardModel, Long>{
 
 	@Query("select a from CreditCardModel a Where a.accountNo=?1")
-	public List<CreditCardModel> getCreditCardDetatils(BigInteger accountNo) ;
+	public List<CreditCardModel> getCreditCardDetatils(long accountNo) ;
 		
 }
